# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "DuAnTotNghiep API",
        "description": "REST API documentation for Users and other resources",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/duan/duan/duantotnghiep/public",
            "description": "Localhost base path"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "2. Users"
                ],
                "summary": "Danh sách người dùng (có phân trang)",
                "description": "API này trả về danh sách người dùng với phân trang và các bộ lọc. Hỗ trợ tìm kiếm theo tên, email, số điện thoại.",
                "operationId": "getUsers",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Tìm kiếm theo tên, email, số điện thoại",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang hiện tại",
                        "schema": {
                            "type": "integer",
                            "minimum": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Số bản ghi mỗi trang",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Trạng thái: 0=không hoạt động, 1=hoạt động",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ]
                        }
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Vai trò: 1=Quản trị, 2=Cộng tác viên",
                        "schema": {
                            "type": "integer",
                            "enum": [
                                1,
                                2
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thành công",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "data": [
                                        {
                                            "id": 1,
                                            "name": "Nguyễn Văn A",
                                            "email": "nguyenvana@example.com",
                                            "phone": "0123456789",
                                            "address": "123 Đường ABC, Quận 1, TP.HCM",
                                            "province_id": 79,
                                            "district_id": 761,
                                            "ward_id": 26782,
                                            "status": 1,
                                            "user_catalogue_id": 2,
                                            "created_at": "2024-01-01T00:00:00.000000Z",
                                            "updated_at": "2024-01-01T00:00:00.000000Z"
                                        }
                                    ],
                                    "pagination": {
                                        "current_page": 1,
                                        "per_page": 10,
                                        "total": 1,
                                        "last_page": 1
                                    },
                                    "message": "Lấy danh sách người dùng thành công"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Lỗi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Có lỗi xảy ra"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "2. Users"
                ],
                "summary": "Tạo mới người dùng",
                "operationId": "88ac51e048bd17c93b767194a54cde27",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "Nguyễn Văn A",
                                "email": "a@example.com",
                                "password": "matkhau@123",
                                "phone": "0901234567",
                                "image": "avatar.jpg",
                                "address": "123 Đường ABC",
                                "province_id": 79,
                                "district_id": 760,
                                "ward_id": 26734,
                                "birthday": "2000-01-01",
                                "description": "Ghi chú",
                                "status": 1,
                                "user_catalogue_id": 2
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tạo mới thành công",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/User"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tạo người dùng thành công"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "Lỗi",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status_code": {
                                            "type": "integer",
                                            "example": 409
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Không thể xóa người dùng vì đang được tham chiếu ở dữ liệu khác (ví dụ: đơn hàng)."
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Nguyễn Văn A"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "a@example.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "0901234567"
                    },
                    "image": {
                        "type": "string",
                        "example": "https://.../avatar.jpg",
                        "nullable": true
                    },
                    "status": {
                        "type": "integer",
                        "enum": [
                            0,
                            1
                        ],
                        "example": 1
                    },
                    "role": {
                        "type": "integer",
                        "enum": [
                            1,
                            2
                        ],
                        "example": 2
                    },
                    "province_id": {
                        "type": "integer",
                        "example": 79,
                        "nullable": true
                    },
                    "district_id": {
                        "type": "integer",
                        "example": 760,
                        "nullable": true
                    },
                    "ward_id": {
                        "type": "integer",
                        "example": 26734,
                        "nullable": true
                    },
                    "address": {
                        "type": "string",
                        "example": "123 ABC, Q1, HCM",
                        "nullable": true
                    },
                    "birthday": {
                        "type": "string",
                        "format": "date",
                        "example": "2000-01-01",
                        "nullable": true
                    },
                    "description": {
                        "type": "string",
                        "example": "Ghi chú",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}